Pull interactions test in as a version-conditional feature
